DJAeun/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langchain-text-splitters>=0.3.0
langsmith>=0.2.0
pinecone-client>=5.0.0
streamlit>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
